Missing imports and a few more functions moving into serialtbx
